added the list of all orders

+++ frontend/src/views/Users.js
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react'
+import React, { useEffect, useState } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { USER_DETAILS_RESET } from '../actions/types'
 import { getUsers, adminDeleteUser } from '../actions/user'
@@ -10,8 +10,9 @@ export default function Users(props) {
     const { error, users, loading } = usersReducer
     const userSignIn = useSelector(state => state.userSignInReducer)
     const { userInfo } = userSignIn
-    const deleteReducer = useSelector(state => state.adminDeleteReducer)
+    //const deleteReducer = useSelector(state => state.adminDeleteReducer)
     //const { errorDelete, successDelete, loadingDelete } = deleteReducer
+    const [userList, setUserList] = useState([])
 
     if (!userInfo) {
         props.history.push("/signin")
@@ -22,14 +23,17 @@ export default function Users(props) {
     useEffect(() => {
         dispatch({ type: USER_DETAILS_RESET })
         dispatch(getUsers())
-    }, [dispatch])
+        if (users) {
+            setUserList(users)
+        }
+    }, [dispatch, userList])
 
     const updateHandler = (userId) => {
         props.history.push(`/profile/${userId}`)
     }
     const deleteHandler = (userId) => {
         dispatch(adminDeleteUser(userId))
-        users.filter(user => user._id !== userId)
+        setUserList(userList.filter(user => user._id !== userId))
     }
 
     return (
@@ -57,7 +61,9 @@ export default function Users(props) {
                                         <button type="button" className="small" onClick={(e) => updateHandler(user._id)}>Edit</button>
                                     </td>
                                     <td>
-                                        <button type="button" className="small" onClick={() => deleteHandler(user._id)}>Delete</button>
+                                        { 
+                                            !user.isAdmin && <button type="button" className="small" onClick={() => deleteHandler(user._id)}>Delete</button>
+                                        }
                                     </td>
                                 </tr>
                             ))}
